browser-emulator: add to package.json (disabled) deletion of DOM types
The DOM types should NOT be enabled on a pure Node.js build. But they
are because of a bug in TypeScript, that allows 3rd-party dependencies
to accidentally pollute the global namespace with its own dependencies.

Canvas depends on DOM, and because of this bug, the whole global
namespace is polluted with DOM types.

This is not helpful when we want to create a clean blank
reimplementation of the WebRTC types (even if they are very similar to
the original ones).