browser-emulator: move video processing scripts from sh to python
This will be used to test if performance is better for video processing, the idea being working with frames in memory instead of multiple calls to ffmpeg. This also allows better multithreading.
Also includes improvements to OCR alignment, using tesseract instead of gocr and retrying with multiple configurations in case of ocr error.