browser-emulator: refactor real browser architecture
Removes selenoid to use only selenium as it is no longer necessary to have an additional layer.
Stop using chrome options for using fake webcam and microphone as it caused desync between them, using instead ffmpeg to feed the fake mediafiles to a pulseaudio virtual microphone and a v4l2loopback virtual camera.